REFACTOR: 메인 인기도서 코드 리팩토링
