--- src/atom/popularList.js
@@ -1,8 +1,0 @@
-import { atom } from "recoil";
-
-const popularList = atom({
-  key: "popularList",
-  default: [],
-});
-
-export default popularList;
