--- src/atom/popularMain.js
@@ -1,20 +1,0 @@
-import { atom } from "recoil";
-
-const popularMain = atom({
-  key: "popularMain",
-  default: {
-    author: [],
-    contents: "",
-    datetime: "",
-    isbn: "",
-    price: 0,
-    publisher: "",
-    sale_price: 0,
-    thumbnail: "",
-    title: "",
-    translators: [],
-    url: "",
-  },
-});
-
-export default popularMain;
